refactor: controller 클래스 분류
