+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,6 +1,6 @@
 package com.techcourse;
 
-import com.techcourse.controller.*;
+import com.techcourse.controller.lagacy.*;
 import jakarta.servlet.http.HttpServletRequest;
 import nextstep.mvc.HandlerMapping;
 import nextstep.mvc.controller.asis.Controller;
@@ -20,15 +20,14 @@ public class ManualHandlerMapping implements HandlerMapping {
     @Override
     public void initialize() {
         controllers.put("/", new ForwardController("/index.jsp"));
-
-                // 어노테이션 기반 컨트롤러로 대체
+        // legacy, 사용을 위해 주석 해제시, annotationcontroller 패키지 내의 컨트롤러 어노테이션 주석 처리 혹은 삭제
 //        controllers.put("/login", new LoginController());
 //        controllers.put("/login/view", new LoginViewController());
 //        controllers.put("/logout", new LogoutController());
 //        controllers.put("/register/view", new RegisterViewController());
 //        controllers.put("/register", new RegisterController());
 
-                log.info("Initialized Handler Mapping!");
+        log.info("Initialized Handler Mapping!");
         controllers.keySet().forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
     }
