+++ app/src/main/java/com/techcourse/controller/annotationcontroller/AuthController.java
@@ -1,5 +1,6 @@
-package com.techcourse.controller;
+package com.techcourse.controller.annotationcontroller;
 
+import com.techcourse.controller.UserSession;
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
