+++ app/src/main/java/com/techcourse/controller/annotationcontroller/MemberController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.annotationcontroller;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
