+++ app/src/main/java/com/techcourse/controller/lagacy/LogoutController.java
@@ -1,5 +1,6 @@
-package com.techcourse.controller;
+package com.techcourse.controller.lagacy;
 
+import com.techcourse.controller.UserSession;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
